move http_client to ddcommon
