+++ crashtracker/src/crash_info/telemetry.rs
@@ -162,7 +162,7 @@ impl TelemetryCrashUploader {
                 count: 1,
             }]),
         };
-        let client = ddtelemetry::worker::http_client::from_config(&self.cfg);
+        let client = ddcommon::http_client::from_endpoint(&self.cfg.endpoint);
         let req = request_builder(&self.cfg)?
             .method(http::Method::POST)
             .header(
